Rel 1.0.0 (#6)
* cleanup

* some bootstrap changes
- add patient document fax listeners

* Revert "some bootstrap changes"

This reverts commit cc1342427f88d1277fea264885751bbede7eeca7.

* add documents listeners

* - refactor

* - update release
- refactor directories to temp

+++ controllers/TwilioFaxClient.php
@@ -8,6 +8,7 @@
  * @copyright Copyright (c) 2019 Jerry Padgett <sjpadgett@gmail.com>
  * @license   https://github.com/openemr/openemr/blob/master/LICENSE GNU General Public License 3
  */
+
 namespace OpenEMR\Modules\FaxSMS\Controllers;
 
 use DateTime;
@@ -30,8 +31,8 @@ class TwilioFaxClient extends AppDispatch
     public function __construct()
     {
         $this->crypto = new CryptoGen();
-        $this->baseDir = $GLOBALS['OE_SITE_DIR'] . "/messageStore";
-        $this->uriDir = $GLOBALS['OE_SITE_WEBROOT'] . "/messageStore";
+        $this->baseDir = $GLOBALS['temporary_files_dir'];
+        $this->uriDir = $GLOBALS['OE_SITE_WEBROOT'];
         $this->cacheDir = $GLOBALS['OE_SITE_DIR'] . '/documents/logs_and_misc/_cache';
         $this->credentials = $this->getCredentials();
         parent::__construct();
@@ -161,7 +162,11 @@ class TwilioFaxClient extends AppDispatch
             file_put_contents($tmpPath, $content);
         }
 
-        $faxfile = $this->uriDir . '/send/' . $basename;
+        // api rest call to fax server to verify twilio request signature and stream file
+        $faxfile = $this->serverUrl . $GLOBALS['webroot'] .
+            '/interface/modules/custom_modules/oe-module-faxsms/faxserver/serveFax?site=' . $this->getSession('site_id') .
+            '&file=' . attr($basename);
+        // callback on completion of fax send
         $callbackUrl = $this->serverUrl . $GLOBALS['webroot'] .
             '/interface/modules/custom_modules/oe-module-faxsms/faxserver/faxCallback?site=' . $this->getSession('site_id');
 
@@ -201,12 +206,6 @@ class TwilioFaxClient extends AppDispatch
             return json_encode($ee);
         };
         try {
-            $messageStoreDir = $this->baseDir;
-
-            //Create the Directory
-            if (!file_exists($messageStoreDir)) {
-                mkdir($messageStoreDir, 0777, true);
-            }
             // dateFrom and dateTo paramteres
             $timeFrom = 'T00:00:01Z';
             $timeTo = 'T23:59:59Z';
@@ -238,7 +237,7 @@ class TwilioFaxClient extends AppDispatch
                     $d2 = new DateTime();
                     $dif = $d1->diff($d2);
                     $interval = ($dif->d * 24) + $dif->h;
-                    if ($interval >= 12 && ($status != 'delivered' || $status != 'received')) {
+                    if ($interval >= 12 && ($status != 'delivered' && $status != 'received')) {
                         $f = $twilio->fax->v1->faxes($id)->delete();
                     } elseif ($interval >= 48) {
                         $f = $twilio->fax->v1->faxes($id)->delete();
